+++ Dracula/ExternalAPIs/IDynaApi.cs
@@ -1,0 +1,13 @@
+﻿using Nickel;
+
+namespace Shockah.Dracula;
+
+public interface IDynaApi
+{
+	IDeckEntry DynaDeck { get; }
+}
+
+public interface IDynaHook
+{
+	void OnChargeTrigger(State state, Combat combat, Ship ship, int worldX) { }
+}