+++ Dracula/docs/release-notes.md
@@ -2,6 +2,10 @@
 
 # Release notes
 
+## Upcoming release
+
+* Added Dracula.EXE.
+
 ## 1.0.4
 Released 29 February 2024.